// packages/fluentui/react/src/themes/types.ts
export type ComponentAreaName =
  | 'foreground'
  | 'background'
  | 'border'
  | 'shadow'
  | 'foregroundHover'
  | 'backgroundHover'
  | 'borderHover'
  | 'shadowHover'
  | 'foregroundActive'
  | 'backgroundActive'
  | 'borderActive'
  | 'shadowActive'
  | 'foregroundFocus'
  | 'backgroundFocus'
  | 'borderFocus'
  | 'shadowFocus'
  | 'foregroundPressed'
  | 'backgroundPressed'
  | 'borderPressed'
  | 'shadowPressed'
  | 'foregroundDisabled'
  | 'backgroundDisabled'
  | 'borderDisabled'
  | 'shadowDisabled';
// ...
export type StrictColorScheme<T extends string | number | symbol = ComponentAreaName> = Record<T, string>;
// ...
export type ItemType<T> = T extends (infer TItem)[] ? TItem : never;